chore: rename to Page
As suggested in PR review, rename to just 'Page'.
