+++ packages/ui/src/lib/index.ts
@@ -23,10 +23,10 @@ import Checkbox from './checkbox/Checkbox.svelte';
 import DropdownMenu from './dropdownMenu';
 import Input from './inputs/Input.svelte';
 import SearchInput from './inputs/SearchInput.svelte';
-import BasePage from './layouts/BasePage.svelte';
 import DetailsPage from './layouts/DetailsPage.svelte';
 import FormPage from './layouts/FormPage.svelte';
 import NavPage from './layouts/NavPage.svelte';
+import Page from './layouts/Page.svelte';
 import Link from './link/Link.svelte';
 import Modal from './modal/Modal.svelte';
 import LinearProgress from './progress/LinearProgress.svelte';
@@ -45,7 +45,6 @@ import { isFontAwesomeIcon } from './utils/icon-utils';
 
 export type { ButtonType };
 export {
-  BasePage,
   Button,
   Checkbox,
   CloseButton,
@@ -60,6 +59,7 @@ export {
   Link,
   Modal,
   NavPage,
+  Page,
   SearchInput,
   SettingsNavItem,
   Spinner,
